package renamed example -> prestashop
