+++ src/main/java/org/prestashop/entities/Customer.java
@@ -1,4 +1,4 @@
-package org.example.entities;
+package org.prestashop.entities;
 
 import com.github.javafaker.Faker;
